requirements.txt
streamlit
faster-whisper
moviepy<2.0.0
google-api-python-client
google-auth-oauthlib
pillow
numpy<2.0.0
